Updated python for native-terraform operation.
